[#59] Feat: 대표배지 관련 기능을 Badge 도메인으로 변경
