+++ src/main/java/com/server/money_touch/domain/badge/controller/BadgeController.java
@@ -1,19 +1,20 @@
 package com.server.money_touch.domain.badge.controller;
 
 import com.server.money_touch.domain.badge.dto.BadgeResponse;
-import com.server.money_touch.domain.badge.service.BadgeService;
+import com.server.money_touch.domain.badge.service.BadgeCommandService;
+import com.server.money_touch.domain.badge.service.BadgeQueryService;
 import com.server.money_touch.global.apiPayload.ApiResponse;
 import com.server.money_touch.global.apiPayload.code.status.ErrorStatus;
 import com.server.money_touch.global.validation.annotation.ApiErrorCodeExample;
 import com.server.money_touch.global.validation.annotation.ApiErrorCodeExamples;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.Parameters;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @Tag(name = "배지 페이지", description = "배지에 관한 API")
 @Slf4j
@@ -23,7 +24,8 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/badge")
 public class BadgeController {
 
-    private final BadgeService badgeService;
+    private final BadgeQueryService badgeQueryService;
+    private final BadgeCommandService badgeCommandService;
 
     @Operation(
             summary = "내가 획득한 배지 목록 조회 API",
@@ -39,7 +41,53 @@ public class BadgeController {
     public ApiResponse<BadgeResponse.MyBadgeListResultDTO> getMyBadges() {
 
         // userId 임시로 1로 지정 (추후 JWT 토큰에서 추출)
-        BadgeResponse.MyBadgeListResultDTO response = badgeService.getMyBadges(1L);
+        BadgeResponse.MyBadgeListResultDTO response = badgeQueryService.getMyBadges(1L);
+        return ApiResponse.onSuccess(response);
+    }
+
+    // 대표 배지 설정
+    @Operation(
+            summary = "대표 배지 설정 API",
+            description = "획득한 배지 중에서 대표 배지를 설정하는 API입니다. UserBadge 테이블에서 획득 여부를 확인 후 User 엔티티의 badgeId를 업데이트합니다. 대표배지는 1개만 설정이 가능하고, 새로 설정하면 교체됩니다. "
+    )
+//    @ApiSuccessCodeExample(resultClass = BadgeResponse.BadgeDetailResultDTO.class)
+    @ApiErrorCodeExamples({
+            @ApiErrorCodeExample(value = ErrorStatus.class, name = "USER_NOT_FOUND"),
+            @ApiErrorCodeExample(value = ErrorStatus.class, name = "BADGE_NOT_FOUND"),
+            @ApiErrorCodeExample(value = ErrorStatus.class, name = "BADGE_NOT_EARNED"),
+            @ApiErrorCodeExample(value = ErrorStatus.class, name = "_BAD_REQUEST"),
+            @ApiErrorCodeExample(value = ErrorStatus.class, name = "_INTERNAL_SERVER_ERROR")
+    })
+    @Parameters({
+            @Parameter(name = "badgeId", description = "설정할 대표 배지 ID", example = "1", required = true)
+    })
+    @PatchMapping("/representative-badge/{badgeId}")
+    public ApiResponse<BadgeResponse.RepresentativeBadgeResultDTO> setRepresentativeBadge(
+//            @AuthenticationPrincipal UserPrincipal userPrincipal,
+            @PathVariable Long badgeId
+    ) {
+        // userId 임시로 1로 지정 (추후 JWT 토큰에서 추출)
+        BadgeResponse.RepresentativeBadgeResultDTO response = badgeCommandService.setRepresentativeBadge(1L, badgeId);
+        return ApiResponse.onSuccess(response);
+    }
+
+    // 대표 배지 조회
+    @Operation(
+            summary = "대표 배지 조회 API",
+            description = "사용자의 대표 배지 정보를 조회하는 API입니다. 대표 배지가 설정되어 있지 않은 경우, 응답은 null입니다."
+    )
+//    @ApiSuccessCodeExample(resultClass = BadgeResponse.BadgeDetailResultDTO.class)
+    @ApiErrorCodeExamples({
+            @ApiErrorCodeExample(value = ErrorStatus.class, name = "USER_NOT_FOUND"),
+            @ApiErrorCodeExample(value = ErrorStatus.class, name = "BADGE_NOT_EARNED"),
+            @ApiErrorCodeExample(value = ErrorStatus.class, name = "_INTERNAL_SERVER_ERROR")
+    })
+    @GetMapping("/representative-badge")
+    public ApiResponse<BadgeResponse.RepresentativeBadgeResultDTO> getRepresentativeBadge(
+//            @AuthenticationPrincipal UserPrincipal userPrincipal
+    ) {
+        // userId 임시로 1로 지정 (추후 JWT 토큰에서 추출)
+        BadgeResponse.RepresentativeBadgeResultDTO response = badgeQueryService.getRepresentativeBadge(1L);
         return ApiResponse.onSuccess(response);
     }
