+++ src/main/java/com/server/money_touch/domain/badge/service/BadgeCommandService.java
@@ -1,0 +1,9 @@
+package com.server.money_touch.domain.badge.service;
+
+import com.server.money_touch.domain.badge.dto.BadgeResponse;
+
+public interface BadgeCommandService {
+
+    // 대표 배지 설정
+    BadgeResponse.RepresentativeBadgeResultDTO setRepresentativeBadge(Long userId, Long badgeId);
+}
