--- src/main/java/com/server/money_touch/domain/user/converter/UserConverter.java
@@ -1,6 +1,0 @@
-package com.server.money_touch.domain.user.converter;
-
-public class UserConverter {
-
-    // Badge 엔터티 -> RepresentativeBadgeResultDTO
-}
