+++ src/main/java/com/server/money_touch/domain/user/dto/UserResponse.java
@@ -62,25 +62,4 @@ public class UserResponse {
         @Schema(description = "대표 뱃지 ID", example = "1")
         private Badge badgeId;
     }
-
-    @Getter
-    @Builder
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Schema(description = "대표 배지 설정 응답")
-    public static class RepresentativeBadgeResultDTO {
-
-        @Schema(description = "설정된 대표 배지 ID", example = "1")
-        private Long badgeId;
-
-        @Schema(description = "배지 이름", example = "절약왕")
-        private String badgeName;
-
-        @Schema(description = "배지 이미지 URL", example = "https://example.com/badge.png")
-        private String badgeImageUrl;
-
-        @Schema(description = "배지 설명", example = "똑똑소비왕 10위권 달성 시")
-        private String badgeDescription;
-
-    }
 }
